Make goblin a proper source module
