+++ barretenberg/cpp/.clangd
@@ -1,5 +1,5 @@
 CompileFlags: # Tweak the parse settings
-  Remove: -fconstexpr-ops-limit=*
+  Remove: -fconstexpr-ops-limit=* -Wfatal-errors
 ---
 # Applies all barretenberg source files
 If:
